move resizeConstraint out of hack file
