+++ src/jsonUtils/hacksForJSONImpl.js
@@ -1,6 +1,5 @@
 // @flow
 // We need native macOS fonts and colors for these hacks so import the old utils
-import type { SJTextStyle } from 'sketchapp-json-flow-types';
 import { toSJSON } from '@skpm/sketchapp-json-plugin';
 import {
   TEXT_ALIGN,
@@ -10,72 +9,8 @@ import {
 } from '../utils/constants';
 import findFont from '../utils/findFont';
 import getSketchVersion from '../utils/getSketchVersion';
-import type { TextNodes, TextNode, TextStyle, ResizeConstraints, LayoutInfo } from '../types';
-import { generateID, makeColorFromCSS } from './models';
-
-/*
-  RESIZE CONSTRAINT RULES
-
-  Order of properties as map keys:
-  1. top
-  2. right
-  3. bottom
-  4: left
-  5. fixedHeight
-  6. fixedWidth
- */
-
-const RESIZE_CONSTRAINTS = {
-  top_left_fixedHeight_fixedWidth: 9,
-  top_right_left_fixedHeight: 10,
-  top_left_fixedHeight: 11,
-  top_right_fixedHeight_fixedWidth: 12,
-  top_fixedHeight_fixedWidth: 13,
-  top_right_fixedHeight: 14,
-  top_fixedHeight: 15,
-  top_bottom_left_fixedWidth: 17,
-  top_right_bottom_left: 18,
-  top_bottom_left: 19,
-  top_right_bottom_fixedWidth: 20,
-  top_bottom_fixedWidth: 21,
-  top_right_bottom: 22,
-  top_bottom: 23,
-  top_left_fixedWidth: 25,
-  top_right_left: 26,
-  top_left: 27,
-  top_right_fixedWidth: 28,
-  top_fixedWidth: 29,
-  top_right: 30,
-  top: 31,
-  bottom_left_fixedHeight_fixedWidth: 33,
-  right_bottom_left_fixedHeight: 34,
-  bottom_left_fixedHeight: 35,
-  right_bottom_fixedHeight_fixedWidth: 36,
-  bottom_fixedHeight_fixedWidth: 37,
-  right_bottom_fixedHeight: 38,
-  bottom_fixedHeight: 39,
-  left_fixedHeight_fixedWidth: 41,
-  right_left_fixedHeight: 42,
-  left_fixedHeight: 43,
-  right_fixedHeight_fixedWidth: 44,
-  fixedHeight_fixedWidth: 45,
-  right_fixedHeight: 46,
-  fixedHeight: 47,
-  bottom_left_fixedWidth: 49,
-  right_bottom_left: 50,
-  bottom_left: 51,
-  right_bottom_fixedWidth: 52,
-  bottom_fixedWidth: 53,
-  right_bottom: 54,
-  bottom: 55,
-  left_fixedWidth: 57,
-  right_left: 58,
-  left: 59,
-  right_fixedWidth: 60,
-  fixedWidth: 61,
-  right: 62,
-  none: 63,
-};
+import type { TextNodes, TextNode, TextStyle, LayoutInfo } from '../types';
+import { makeColorFromCSS } from './models';
 
 // NOTE(gold): toSJSON doesn't recursively parse JS objects
 // https://github.com/airbnb/react-sketchapp/pull/73#discussion_r108529703
@@ -135,50 +70,6 @@ export const makeImageDataFromUrl = (url: string): MSImageData => {
   }
   return MSImageData.alloc().initWithImage(image);
 };
-
-export function makeResizeConstraint(resizingConstraint: ?ResizeConstraints): number {
-  if (resizingConstraint) {
-    const constraints = [];
-    const {
-      top, right, bottom, left, fixedHeight, fixedWidth,
-    } = resizingConstraint;
-
-    if (top) {
-      constraints.push('top');
-    }
-    if (right) {
-      constraints.push('right');
-    }
-    if (bottom) {
-      constraints.push('bottom');
-    }
-    if (left) {
-      constraints.push('left');
-    }
-    if (fixedHeight) {
-      constraints.push('fixedHeight');
-    }
-    if (fixedWidth) {
-      constraints.push('fixedWidth');
-    }
-
-    if (constraints.length > 0) {
-      const constraint = RESIZE_CONSTRAINTS[constraints.join('_')];
-      if (!constraint) {
-        throw new Error(
-          `\n${JSON.stringify(
-            resizingConstraint,
-            null,
-            2,
-          )}\nconstraint is not a valid combination.`,
-        );
-      }
-      return constraint;
-    }
-  }
-
-  return RESIZE_CONSTRAINTS.none; // No constraints
-}
 
 // This shouldn't need to call into Sketch, but it does currently, which is bad for perf :(
 function createStringAttributes(textStyles: TextStyle): Object {
@@ -240,40 +131,28 @@ export function makeEncodedAttributedString(textNodes: TextNodes) {
   return encodeSketchJSON(msAttribStr);
 }
 
-export function makeTextStyle(textStyle: TextStyle) {
+export function makeEncodedTextStyleAttributes(textStyle: TextStyle) {
   const pStyle = makeParagraphStyle(textStyle);
 
   const font = findFont(textStyle);
 
   const color = makeColorFromCSS(textStyle.color || 'black');
 
-  const value: SJTextStyle = {
-    _class: 'textStyle',
-    encodedAttributes: {
-      MSAttributedStringFontAttribute: encodeSketchJSON(font.fontDescriptor()),
-      NSFont: font,
-      NSColor: encodeSketchJSON(
-        NSColor.colorWithDeviceRed_green_blue_alpha(
-          color.red,
-          color.green,
-          color.blue,
-          color.alpha,
-        ),
-      ),
-      NSParagraphStyle: encodeSketchJSON(pStyle),
-      NSKern: textStyle.letterSpacing || 0,
-      MSAttributedStringTextTransformAttribute:
-        TEXT_TRANSFORM[textStyle.textTransform || 'initial'] * 1,
-    },
-  };
-
   return {
-    _class: 'style',
-    sharedObjectID: generateID(),
-    miterLimit: 10,
-    startDecorationType: 0,
-    endDecorationType: 0,
-    textStyle: value,
+    MSAttributedStringFontAttribute: encodeSketchJSON(font.fontDescriptor()),
+    NSFont: font,
+    NSColor: encodeSketchJSON(
+      NSColor.colorWithDeviceRed_green_blue_alpha(
+        color.red,
+        color.green,
+        color.blue,
+        color.alpha,
+      ),
+    ),
+    NSParagraphStyle: encodeSketchJSON(pStyle),
+    NSKern: textStyle.letterSpacing || 0,
+    MSAttributedStringTextTransformAttribute:
+      TEXT_TRANSFORM[textStyle.textTransform || 'initial'] * 1,
   };
 }
